Original file for my Fantasy Baseball program
A short Fantasy Baseball program written in 2010 that is used to easily
take a list of player names and put them in a different order.  It also
removed BATTER and PITCHER from the lines.  My reasoning for creating
this 5 years ago is beyond my 2015 comprehension, so this is just a
Sandbox test to see that I can commit.